# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buleje / Bodega San Martín API",
    "version": "0.1.0",
    "description": "ERP/e-commerce multi-tenant API para Buleje — Pucallpa, Perú. Endpoints de /admin requieren autenticación JWT con rol válido."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "CouponValidResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "couponId": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "discountType": {
            "type": "string",
            "enum": [
              "percent",
              "fixed"
            ]
          },
          "discountValue": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          }
        },
        "required": [
          "valid",
          "couponId",
          "code",
          "description",
          "discountType",
          "discountValue",
          "discount"
        ]
      },
      "CouponInvalidResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "reason": {
            "type": "string",
            "example": "Cupón expirado"
          }
        },
        "required": [
          "valid",
          "reason"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "CouponValidateBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "example": "BIENVENIDO10"
          },
          "storeSlug": {
            "type": "string",
            "minLength": 1,
            "example": "bodega-san-martin"
          },
          "cartTotal": {
            "type": "number",
            "exclusiveMinimum": 0,
            "example": 45.5
          }
        },
        "required": [
          "code",
          "storeSlug",
          "cartTotal"
        ]
      },
...
      "StoreRegisterResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "isPublished": {
                "type": "boolean"
              },
              "tenantId": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "slug",
              "name",
              "isPublished",
              "tenantId"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "StoreRegisterBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Bodega Las Flores"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "example": "Bodega familiar con productos frescos"
          },
          "logo": {
            "type": "string",
            "format": "uri"
          },
          "banner": {
            "type": "string",
            "format": "uri"
          },
          "category": {
            "type": "string",
            "maxLength": 50,
            "example": "abarrotes"
          },
          "zone": {
            "type": "string",
            "maxLength": 80,
            "example": "Pucallpa Centro"
          },
          "commission": {
            "type": "number",
            "minimum": 0,
            "maximum": 30,
            "example": 10
          }
        },
        "required": [
          "name"
        ]
      },
      "DriverApplyResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "DriverApplyBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "Juan Pérez"
          },
          "phone": {
            "type": "string",
            "minLength": 6,
            "maxLength": 20,
            "example": "51987654321"
          },
          "zone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "Pucallpa Centro"
          },
          "vehicleType": {
            "type": "string",
            "enum": [
              "moto",
              "bicicleta",
              "auto",
              "a_pie"
            ],
            "example": "moto"
          },
          "availability": {
            "type": "string",
            "enum": [
              "manana",
              "tarde",
              "noche",
              "full",
              "fines"
            ],
            "example": "full"
          }
        },
        "required": [
          "name",
          "phone",
          "zone",
          "vehicleType",
          "availability"
        ]
      },
      "QAListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QAItem"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "QAItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userName": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userName",
          "question",
          "answer",
          "createdAt"
        ]
      },
      "QACreatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QAItem"
          }
        },
        "required": [
          "data"
        ]
      },
      "QAQuestionBody": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "example": "María López"
          },
          "question": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500,
            "example": "¿Este producto tiene garantía?"
          }
        },
        "required": [
          "userName",
          "question"
        ]
      },
...
    },
...
  },
  "paths": {
    "/marketplace/coupons/validate": {
      "post": {
        "summary": "Validar cupón de descuento",
        "tags": [
          "Marketplace",
          "Coupons"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CouponValidateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado de validación (válido o inválido con razón)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CouponValidResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CouponInvalidResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tienda no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/marketplace/stores/register": {
      "post": {
        "summary": "Registrar nueva tienda en el marketplace",
        "tags": [
          "Marketplace",
          "Stores"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreRegisterBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tienda registrada (pendiente de aprobación)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Requiere rol admin u owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Tenant ya tiene tienda registrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/drivers/apply": {
      "post": {
        "summary": "Postular como repartidor",
        "tags": [
          "Marketplace",
          "Drivers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverApplyBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud registrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverApplyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error de servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/qa/{productId}": {
      "get": {
        "summary": "Listar preguntas y respuestas de un producto",
        "tags": [
          "Marketplace",
          "QA"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "42",
              "description": "ID numérico del producto"
            },
            "required": true,
            "description": "ID numérico del producto",
            "name": "productId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de preguntas del producto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAListResponse"
                }
              }
            }
          },
          "400": {
            "description": "productId inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Registrar nueva pregunta sobre un producto",
        "tags": [
          "Marketplace",
          "QA"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "42",
              "description": "ID numérico del producto"
            },
            "required": true,
            "description": "ID numérico del producto",
            "name": "productId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QAQuestionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pregunta registrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QACreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o ID inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}